Implement quorum reads and merging for /v2/silences and /v2/silence/{id}. (#4141)
Reading silence listings and individual silences will now read from a
quorum of replicas and return a merged response. In both cases, in the
presence of duplicate silences with the same "id", the silence with the
most recent "updatedAt" timestamp is return.

Note that this does not yet bring full consistency to silences, as we
are not able to perform quorum writing of silences without upstream
changes to Alertmanager.

However, it will still be an improvement:
- Reads will be consistent more often, being able to take state from
  multiple replicas instead of just one.
- Requests will be resilient to single replica failure, as the request
  is essentially attempted on all three replicas.

An possible extension to this change would be to block for all replicas
to reply, but still allow a single failure. This would mean that reads
are consistent for the case when all replicas are contactable.

Signed-off-by: Steve Simpson <steve.simpson@grafana.com>
Former-commit-id: 653affe7eea1596335f9d53e07189e88bace1b4b