HA-Tracker: do KV-Store updates in the background (#271)
* HA-Tracker: do KV-Store updates in the background

We don't want to hold up normal flow of push requests while we do a
compare-and-swap on KVStore, which can take hundreds of milliseconds.

Instead, store the last-seen timestamp of both the elected replica and
any other replica we see (presumably its fail-over pair), then update
KVStore periodically in the background.

In general we will only update the cache via a callback from the kv store.
One exception is if we are adding the very first record for a cluster,
as part of a push operation. Now we lock around that operation so we
don't overwrite an update from the kv callback.