Fix the integration test workflow to actually use Go 1.17.3 (#543)
The commands to download and install the newer version of Go for running the
integration tests do not work, due to Go actually being installed in a
different location to `/usr/local`, the binary actually used to run the
tests is `/opt/hostedtoolcache/go/1.15.15/x64/bin/go`.

For added confusion, this is not a stock Ubuntu 20.04, but the official GitHub
image with some tools pre-installed or pre-cached. Therefore, it would be
safer to use the officially supported method of updating Go in this container,
which is `actions/setup-go`.