Alertmanager: Remove ability to run without clustering. (#1044)
This change removes the ability for Alertmanager to run without clustering
enabled if sharding is not enabled. Tests have been updated to run with
sharding enabled where necessary, though I have decided against a full
clean-up of the tests to avoid polluting the significant test changes.

To keep this change small and easier to review, a follow up commit will remove
the ability to run without sharding enabled.