Mixin: Show in-memory series and the per-user limit on Tenants dashboard. (#1613)
Changes the "Active series" panel on the "Tenants" dashboard to show:

- Current number of in-memory series for the tenant
- The current in-memory series limit (either from overrides or the default)

This information is extremely valuable to operators tasked with troubleshooting when tenants are hitting limits, or whether they are close to hitting limits. Looking at only active series often causes confusion because limits are applied to the number of in-memory series, not active series.