Allow custom data source regex in mixin dashboards (#1802)
* dashboards: update grafana-builder

The following commit update grafana-builder version and brings in:
* enable toolip by default (#665)
* Add 'Data Source' label for the default datasource template variable. (#672)
* add dashboard link func (#683)
* make allValue configurable (#703)
* Allow datasource's regex to be configured

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>

* Allow custom data source regex in mixin dashboards

The current dashboards offer the possibility to select a data source
among all prometheus data sources in the organization. Depending on the
number of data sources the list could be rather big (>10). Not all data
sources host Mimir metrics as such listing them is not helpful for the
users.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>

* Revert back change that was enabling shared tooltips

Signed-off-by: Marco Pracucci <marco@pracucci.com>

Co-authored-by: Marco Pracucci <marco@pracucci.com>