Make networking panels pod matchers work with helm chart (#1927)
* Make networking panels pod matchers work with helm chart

The pods created by the helm chart follow a format of
<helm_release_name>-mimir-<ingester|distributor|...>.

This is a problem for all places that use the per_instance_label for
matching. The per_instance_label is mostly used in aggregations (sum by
(pod), count by (pod), ...). The networking panels are the only ones
that use it for matching.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Replace .* with a stronger regex in pod matchers

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Add CHANGELOG.md entry

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>