Report previously failed shipped blocks in shipper (#2435)
* Report previously failed shipped blocks in shipper

When a block upload appears to have failed, but later the shipper
discovers that the upload actually succeeded, the shipper does not
upload it again. But the ingester still reports the block as unhsipped.
This PR changes the shipper to report an upload when it discovers
such a block. This helps with invalidating the cached shipper meta file
and reporting correct metrics.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Add CHANGELOG.md entry

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Add tests for false negative meta file uploads

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Use stdErr as a logger

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>