Add scale down policies for autoscaling (#2962)
* Add scale down policies for autoscaling

[By default](https://kubernetes.io/docs/tasks/run-application/horizontal-pod-autoscale/#default-behavior),
autoscaling allows to scale down to minReplicas in one go. This is
problematic in cases where the external metric source for autoscaling is
 unavailable for longer than the stabilization window for scaling down.
That window is 5m by default.

If the scaling metric needs some time to start increasing (if for
example it's using sum(rate()) over longer periods), this scale down
will take longer to recover from. This PR allows to scale down only 10%
of replicas every 1m. This will give more time for the metric source to
scrape metrics and for the scaling metric to stabilize before we reach
minReplicas.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>