Store-gateway: shortcut some cache and bucket ops when matching 0 series (#3848)
In the case some matchers in the request try to select a non-existent label value this PR short-circuits the lookup in the cache and the bucket by using the index header first.

Why? Because sometimes the list of postings can be massive (up to 1MB of diff-encoded, snappy-compressed postings). Fetching and decoding these from the cache or the bucket takes a lot of effort for something we can tell by doing a lookup on the local disk.

This comes with the drawback that now we will do more lookups to disk. With the streaming binary reader this should not be a problem, but with the mmap binary reader this may lead to more frequent mmap page faults.

There is the detail that we also do a lookup in the index header on a cache miss (here), which is duplicated with