store-gateway: add partitioners for each data type (#4095)
* store-gateway: add partitioners for each data type

The store-gateway uses a partitioner to plan the bytes to fetch from the
 bucket. This partitioner is responsible for merging byte ranges that
 overlap and byte ranges that are close-by (<512 KiB apart).

The same partitioner instance is used for partitioning chunks, series,
and postings byte ranges. The default gap size of 512KiB might have a
different effect on the different data types. For example for postings,
512KiB might be well over the total size of all posting lists. Whereas
for chunks it might be insignificant.

This makes it difficult to work with the metrics that the
partitioner exposes - bytes requested (sum of all byte ranges before
merging overlapping and extending close-by ranges) vs bytes
expanded (sum of all byte ranges after merging overlapping and extending
close-by ranges).

This PR creates 3 partitioners to use for each data_type. Each
partitioner has its own metrics with the `data_type` label. The values
for the label are `chunks`, `series`, `postings`.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Add changelog entry

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Rename args and unexport fields

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>