Helm: Multiple topology spread constraints support (#4584)
We have a use case where we need the ability to define multiple topology
spread constraints. Kubernetes supports this as documented under
https://kubernetes.io/docs/concepts/scheduling-eviction/topology-spread-constraints/.

Before this PR we only supported on topology spread constraint and this
now adds support for multiple ones. Backwards compatibility is kept by
supporting the old object/dict syntax next to the new array based syntax
as shown in test-oss-topology-spread-constraints-values.yaml.