Add nginx to dev docker-compose stacks (#4877)
* Add nginx to dev docker-compose stacks

This change adds nginx support in front of the microservices and read-write mode dev docker-compose stacks. This allows all API requests against a single datasource in Grafana to be routed to the appropriate services. The motivation for this is being able to view/create/edit Mimir backed alert rules in Grafana, which only works when using a reverse proxy, and with these changes is supported out of the box. The pre-defined alert rules in mimir-microservices-mode will now show up when starting up.

* Update nginx port to 8080

* Update docker config ports for nginx

* Use a single nginx.conf.template to render nginx.conf from ENV vars