Upgrade prometheus/alertmanager to latest (#4882)
This PR upgrades the prometheus/alertmanager dependency. This is primarily to support future work to solve #4200, which relies on recent changes to the AM's template package. However, it also incorporates a few fixes and extensions to notifiers.

This change also disallows the new configuration options `telegram bot_token_file` and `webhook url_file`.