store-gateway: more precise estimations of LabelValues size (#4844)
* store-gateway: more precise estimations of LabelValues size

When there is a prefix in the matcher, such as `pod=~"querier.*"`, the postings offset table is optimized to only scan
the values that match this prefix. This reduces the duration of the whole lookup. PostingsOffsets also tries to
preallocate the size of the slice with matching values. When we have a prefix we can be more precise in estimating where
is the last value which matches the prefix.

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Co-authored-by: Marco Pracucci <marco@pracucci.com>