store-gateway: stop relying on fixed number of samples per chunk (#5946)
* store-gateway: stop relying on fixed number of samples per chunk

This is a prerequisite to PR 5927

Number of samples per chunk is changing in prometheus. Currently, only for native histograms, but also regular XOR chunks are capped at 1KiB.

This PR changes some tests to not rely on a hardcoded number of samples and also updates/removes some constants in the code. For chunk estimations I changed the tests to assert that the estimation is not off by more than 10% from the real number of chunks.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Compare estimate with returned chunks instead of chunks on disk

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>