Make align-queries-with-step a per-tenant setting (#6714)
Allow the start and end times of queries to be adjusted to be step-aligned
at a per-tenant level. Adjusting the start and end times to multiples of
the step allows for better caching in exchange for behavior that isn't allowed
by Prometheus compatibility testing. Some tenants are OK with that tradeoff
but we were previously unable to set at the tenant level.

Fixes #6712

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>