ruler: add more details to failures for TestRulerEvaluationDelay (#7217)
Related to XXX

This test is flaky and it's always flaky with the same failure - the first sample's value is 4 instead of 2. My hypothesis is that when `now.Add(-evaluationDelay)` is done, then the time range `[now.Add(-evaluationDelay), now]` might become so very slightly smaller than `evaluationDelay` when taken with the `1s` step. This results in sending a query which doesn't quite include the `5m` of samples, but `4.99999(9)m` of samples, which excludes the first sample with value `2`

I couldn't get the test to fail locally, so I'm changing the assertion type to not abort the test on the first mismatch so that we can investigate the all returned samples. Also some debug logs to verify this hypothesis with the step alignment.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>