Add utf8-strict-mode flag for mimirtool (#7227)
* Add utf8-strict-mode flag for mimirtool

This commit adds a new flag to mimirtool called utf8-strict-mode.
It is the client version the utf8-strict-mode-enabled flag in Mimir.

The default mode of mimirtool is to first use the new parser for
UTF-8 matchers, and if it fails, fallback to the classic parser.
If this happens, it also logs a warning to stdout.

It is possible to disable the fallback, and use just the UTF-8 parser,
with the -utf8-strict-mode flag. This will help operators ensure their
configurations are compatible with the new parser going forward.