store-gateway: fix flaky TestBucketStore_Series_TimeoutGate (#8101)
* store-gateway: fix flaky TestBucketStore_Series_TimeoutGate

charleskorn mentioned this. The problem was that the first call in the goroutine wasn't blocking when the client doesn't read the response. Even though responses are streaming there are buffers in gRPC, so the store-gateway can send the response without having to wait for the client to receive it.

After increasing the size of the response, so that the buffers aren't enough, the test didn't terminate. The reason was that we were trying to gracefully shut down the server before cancelling the client request first. This was because of the order of t.Cleanup's and defers. I changed the usages from t.Cleanup to defers to be more consistent with the rest of the file and also reordered the init steps so that they work with both defers and t.Cleanup's.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Pull in Krajo's change

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Co-authored-by: George Krajcsovits <krajorama@users.noreply.github.com>
Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Co-authored-by: George Krajcsovits <krajorama@users.noreply.github.com>